GR initial commit
Written code in python (to make diff easier), rather than an ipython notebook. I have updated the ipynb if there is an easy way to diff with that.

Made a pipeline for submission to scale features and a gridsearch to look over best learning parameters.

Split the data into CV and train with a CV scorer working on the mean squared error for some quantification of each model.

Added submission of final results to CSV.

Plotted the feature correlations as a start to the EDA to work out how we can work with features - it may be best to join longitude and latitude somehow due to their strong correlation.